Refresh meta files following migration to Media WG
This updates the `CONTRIBUTING.md` to explain new contributing conditions
following migration to the Media Working Group.

This also drops the contributing bits from the `LICENSE.md` file, and update
the `w3c.json` file to target the Media Working Group and not the WICG anymore.

README now also mentions the group and links to the explainer.