Drop code that would publish to /TR
While it won't try to publish the spec to /TR in the absence of the secret
token, the spec-prod action still tries to generate the working draft, which
currently fails because there is no /TR link.

Removing the W3C build related options for now, which should (hopefully) make
the spec-prod action stop generating the WD internally.